alert for invalid UUID
